added contractor badge to expenses table

+++ database/migrations/2017_06_24_163231_create_role_user_pivot_table.php
@@ -13,15 +13,17 @@ class CreateRoleUserPivotTable extends Migration
      */
     public function up()
     {
-        Schema::create('role_user', function (Blueprint $table) {
-            $table->integer('role_id')->unsigned();
-            $table->integer('user_id')->unsigned();
+        if (!Schema::hasTable('role_user')) {
+            Schema::create('role_user', function (Blueprint $table) {
+                $table->integer('role_id')->unsigned();
+                $table->integer('user_id')->unsigned();
 
-            $table->foreign('role_id')->references('id')->on('roles')->onDelete('cascade');
-            $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
+                $table->foreign('role_id')->references('id')->on('roles')->onDelete('cascade');
+                $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
 
-            $table->primary(['user_id', 'role_id']);
-        });
+                $table->primary(['user_id', 'role_id']);
+            });
+        }
     }
 
     /**
